SaveScreen refactoring.
