+++ src/routes/index.tsx
@@ -93,7 +93,7 @@ const Routes: React.FC = () => {
 					component={TransactionsScreen}
 					options={(props) => defaultOptions(props)}
 				/>
-				<Stack.Screen name="Save" component={SaveScreen} />
+				<Stack.Screen name="SaveScreen" component={SaveScreen} />
 				<Stack.Screen name="Deposit" component={DepositScreen} />
 				<Stack.Screen name="USDCoin" component={USDCoinScreen} />
 				<Stack.Screen name="DepositSuccess" component={DepositSuccessScreen} />
