+++ src/screens/AssetsScreen/Chart/Header/Header.styles.tsx
@@ -3,5 +3,14 @@ import { StyleSheet } from 'react-native';
 export default StyleSheet.create({
 	container: {
 		justifyContent: 'space-between'
+	},
+	titleContainer: {
+		justifyContent: 'center',
+		alignItems: 'center'
+	},
+	indicatorContainer: {
+		flexDirection: 'row',
+		alignItems: 'center',
+		borderWidth: 3
 	}
 });
