+++ src/screens/SaveScreen/Body/Body.types.ts
@@ -1,0 +1,5 @@
+import { AaveBalances } from '@models/deposit';
+
+export interface BodyProps {
+	lending: AaveBalances['products']['products'][0];
+}
