+++ src/screens/SaveScreen/CurrentValue/CurrentValue.types.ts
@@ -1,0 +1,6 @@
+import { AaveMarket } from '@models/deposit';
+
+export interface CurrentValueProps {
+	depositsBalance: number;
+	aaveMarket: AaveMarket;
+}
