+++ src/screens/SaveScreen/Header/Header.styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	headerNavigation: {
+		marginTop: 12,
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+		alignItems: 'center'
+	},
+	headerNavigationLeft: {
+		flexDirection: 'row',
+		alignItems: 'center'
+	}
+});
