+++ src/screens/WalletScreen/components/index.ts
@@ -1,7 +1,5 @@
 import AssetsPanel from './AssetsPanel/AssetsPanel';
 import ActionsPanel from './ActionsPanel/ActionsPanel';
 import Header from './Header/Header';
-import Accounts from './Accounts/Accounts';
-import Transactions from './Transactions/Transactions';
 
-export { AssetsPanel, ActionsPanel, Header, Accounts, Transactions };
+export { AssetsPanel, ActionsPanel, Header };
