+++ src/screens/WalletScreen/screens/Accounts/Accounts.styles.ts
