+++ src/screens/WalletScreen/screens/Accounts/Image/Image.styles.ts
