+++ src/screens/WalletScreen/screens/Accounts/Image/Image.types.ts
