+++ src/screens/WalletScreen/screens/Transactions/TransactionIcon/TransactionIcon.styles.ts
