+++ src/screens/WalletScreen/screens/Transactions/TransactionIcon/TransactionIcon.tsx
