+++ src/screens/WalletScreen/screens/index.ts
@@ -1,0 +1,4 @@
+import Transactions from './Transactions/Transactions';
+import Accounts from './Accounts/Accounts';
+
+export { Transactions, Accounts };
